Handle (system-wide) proxies better
This pulls in a package that basically asks the embedded Chromium to
resolve any proxies we require, which means we end up using the system
settings.
