+++ background.ts
@@ -6,8 +6,6 @@ import { URL } from 'url';
 
 import Electron from 'electron';
 import _ from 'lodash';
-import MacCA from 'mac-ca';
-import WinCA from 'win-ca';
 
 import mainEvents from '@/main/mainEvents';
 import { setupKim } from '@/main/kim';
@@ -18,6 +16,7 @@ import * as K8s from '@/k8s-engine/k8s';
 import resources from '@/resources';
 import Logging from '@/utils/logging';
 import * as childProcess from '@/utils/childProcess';
+import setupNetworking from '@/main/networking';
 import setupUpdate from '@/main/update';
 
 Electron.app.setName('Rancher Desktop');
@@ -55,10 +54,9 @@ Electron.app.whenReady().then(async() => {
   } catch (err) {
     console.log(`Can't get app version: ${ err }`);
   }
-  if (os.platform().startsWith('win')) {
-    // Inject the Windows certs.
-    WinCA({ inject: '+' });
-  }
+
+  setupNetworking();
+
   try {
     tray = new Tray();
   } catch (e) {
@@ -210,52 +208,6 @@ Electron.ipcMain.handle('settings-write', (event, arg: Partial<settings.Settings
   tray?.emit('settings-update', cfg);
 
   Electron.ipcMain.emit('k8s-restart-required');
-});
-
-// Set up certificate handling for system certificates on Windows and macOS
-Electron.app.on('certificate-error', (event, webContents, url, error, certificate, callback) => {
-  if (error === 'net::ERR_CERT_INVALID') {
-    // If we're getting *this* particular error, it means it's an untrusted cert.
-    // Ask the system store.
-    console.log(`Attempting to check system certificates for ${ url } (${ certificate.subjectName }/${ certificate.fingerprint })`);
-    if (os.platform().startsWith('win')) {
-      const certs: string[] = [];
-
-      WinCA({
-        format: WinCA.der2.pem, ondata: certs, fallback: false
-      });
-      for (const cert of certs) {
-        // For now, just check that the PEM data matches exactly; this is
-        // probably a little more strict than necessary, but avoids issues like
-        // an attacker generating a cert with the same serial.
-        if (cert === certificate.data) {
-          console.log(`Accepting system certificate for ${ certificate.subjectName } (${ certificate.fingerprint })`);
-          // eslint-disable-next-line node/no-callback-literal
-          callback(true);
-
-          return;
-        }
-      }
-    } else if (os.platform() === 'darwin') {
-      for (const cert of MacCA.all(MacCA.der2.pem)) {
-        // For now, just check that the PEM data matches exactly; this is
-        // probably a little more strict than necessary, but avoids issues like
-        // an attacker generating a cert with the same serial.
-        if (cert === certificate.data) {
-          console.log(`Accepting system certificate for ${ certificate.subjectName } (${ certificate.fingerprint })`);
-          // eslint-disable-next-line node/no-callback-literal
-          callback(true);
-
-          return;
-        }
-      }
-    }
-  }
-
-  console.log(`Not handling certificate error ${ error } for ${ url }`);
-
-  // eslint-disable-next-line node/no-callback-literal
-  callback(false);
 });
 
 Electron.ipcMain.on('k8s-state', (event) => {
