+++ package.json
@@ -33,6 +33,7 @@
     "cron-validator": "^1.2.1",
     "cronstrue": "^1.114.0",
     "dompurify": "^2.2.9",
+    "electron-proxy-agent": "^1.2.0",
     "electron-updater": "^4.3.9",
     "fs-extra": "^10.0.0",
     "intl-messageformat": "^7.8.4",
