+++ src/typings/electron-proxy-agent.d.ts
@@ -1,0 +1,11 @@
+declare module 'electron-proxy-agent' {
+  import { Agent } from 'https';
+  import Electron from 'electron';
+
+  export default class ElectronProxyAgent extends Agent {
+    constructor(session?: Electron.Session);
+
+    /** Internally required by http/https client !? */
+    protocol?: string;
+  }
+}
